pyproject.toml
[tool.poetry]
name = "goodreads-to-kindle"
version = "0.1.0"
description = ""
authors = ["Pietro Fanti <pietrofanti.14@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dateutil = "^2.9.0.post0"
scrapy = "^2.12.0"
yapf = "^0.43.0"
rich = "^13.9.4"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.8.1"
zlibrary = "^1.0.1"
dataclasses-json = "^0.6.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
